Fix the response XML for the MiniServer and improve updates when the database is disabled.
git-svn-id: svn://svn.jbidwatcher.com/jbidwatcher/trunk@677 b1acfa68-eb39-11db-b167-a3a8cd6b847e

+++ src/com/jbidwatcher/auction/AuctionEntry.java
@@ -1612,6 +1612,8 @@ public class AuctionEntry extends ActiveRecord implements Comparable {
       //  and the IDs are different, delete the old one.
       if (oldAuction != null &&
           oldAuction != mAuction &&
+          mAuction.getId() != null &&
+          oldAuction.getId() != null &&
           !mAuction.getId().equals(oldAuction.getId())) {
         oldAuction.delete();
       }
